+++ code/src/cljs/sixsq/nuvla/ui/edges/events.cljs
@@ -97,37 +97,35 @@
                     :type    :error})])
       {:db (assoc db ::spec/nuvlaboxes nuvlaboxes
                      ::main-spec/loading? false)
-       :fx [[:dispatch [::get-nuvlaboxes-next-heartbeats nuvlaboxes]]]})))
+       :fx [[:dispatch [::get-nuvlaedges-status nuvlaboxes]]]})))
 
 
 (reg-event-fx
-  ::get-nuvlaboxes-next-heartbeats
+  ::get-nuvlaedges-status
   (fn [_ [_ {nuvlaboxes :resources}]]
     (when (seq nuvlaboxes)
       {::cimi-api-fx/search
        [:nuvlabox-status
-        {:select "parent,next-heartbeat,id"
-         :filter (general-utils/join-and
-                   "online=false"
-                   (apply general-utils/join-or
-                          (map #(str "parent='" (:id %) "'") nuvlaboxes)))}
-        #(dispatch [::set-nuvlaboxes-next-heartbeats %])]})))
+        {:select "parent,next-heartbeat,id,nuvlabox-engine-version,online"
+         :filter (apply general-utils/join-or
+                        (map #(str "parent='" (:id %) "'") nuvlaboxes))}
+        #(dispatch [::set-nuvlaedges-status %])]})))
 
 
 (reg-event-fx
-  ::set-nuvlaboxes-next-heartbeats
+  ::set-nuvlaedges-status
   (fn [{:keys [db]} [_ {:keys [resources] :as nuvlaboxes-status}]]
     (if (instance? js/Error nuvlaboxes-status)
-      (dispatch [::messages-events/add
-                 (let [{:keys [status message]} (response/parse-ex-info nuvlaboxes-status)]
-                   {:header  (cond-> (str "failure getting last online status of offline nuvlaboxes")
-                                     status (str " (" status ")"))
-                    :content message
-                    :type    :error})])
-      {:db (assoc db ::spec/next-heartbeats-offline-edges (zipmap
-                                                            (map :parent resources)
-                                                            (map :next-heartbeat resources))
-                     ::main-spec/loading? false)})))
+      {:fx [[:dispatch [::messages-events/add
+                                 (let [{:keys [status message]} (response/parse-ex-info nuvlaboxes-status)]
+                                   {:header  (cond-> (str "failure getting status for nuvla edges")
+                                               status (str " (" status ")"))
+                                    :content message
+                                    :type    :error})]]]}
+      {:db (assoc db ::spec/nuvlaedges-select-status (zipmap
+                                                      (map :parent resources)
+                                                      resources)
+                  ::main-spec/loading? false)})))
 
 (reg-event-fx
   ::get-nuvlabox-locations