+++ code/src/cljs/sixsq/nuvla/ui/edges/views.cljs
@@ -18,6 +18,7 @@
     [sixsq.nuvla.ui.panel :as panel]
     [sixsq.nuvla.ui.plugins.full-text-search :as full-text-search-plugin]
     [sixsq.nuvla.ui.plugins.pagination :as pagination-plugin]
+    [sixsq.nuvla.ui.session.subs :as session-subs]
     [sixsq.nuvla.ui.utils.forms :as utils-forms]
     [sixsq.nuvla.ui.utils.general :as general-utils]
     [sixsq.nuvla.ui.utils.map :as map]
@@ -26,7 +27,8 @@
     [sixsq.nuvla.ui.utils.style :as style]
     [sixsq.nuvla.ui.utils.ui-callback :as ui-callback]
     [sixsq.nuvla.ui.utils.values :as values]
-    [sixsq.nuvla.ui.utils.zip :as zip]))
+    [sixsq.nuvla.ui.utils.zip :as zip]
+    [sixsq.nuvla.ui.utils.form-fields :as ff]))
 
 
 (def view-type (r/atom :cards))
@@ -233,7 +235,7 @@
                   [:span (str (@tr [:nuvlabox-playbooks-cronjob-ready])
                               " ")
                    (values/copy-value-to-clipboard
-                    "" (@playbooks-cronjob :cronjob) (@tr [:copy-to-clipboard]) true)]
+                     "" (@playbooks-cronjob :cronjob) (@tr [:copy-to-clipboard]) true)]
                   (@tr [:nuvlabox-playbooks-cronjob-wait]))]])
 
             (when @new-private-ssh-key
@@ -732,10 +734,12 @@
 
 
 (defn NuvlaboxRow
-  [{:keys [id name description created state tags online refresh-interval] :as _nuvlabox} managers]
+  [{:keys [id name description created state tags online refresh-interval version created-by] :as _nuvlabox} managers]
   (let [uuid                  (general-utils/id->uuid id)
         locale                (subscribe [::i18n-subs/locale])
-        next-heartbeat-moment @(subscribe [::subs/next-heartbeat-moment id])]
+        next-heartbeat-moment @(subscribe [::subs/next-heartbeat-moment id])
+        engine-version        @(subscribe [::subs/engine-version id])
+        creator               (subscribe [::session-subs/resolve-user created-by])]
     [ui/TableRow {:on-click #(dispatch [::history-events/navigate (str "edges/" uuid)])
                   :style    {:cursor "pointer"}}
      [ui/TableCell {:collapsing true}
@@ -745,7 +749,9 @@
      [ui/TableCell (or name uuid)]
      [ui/TableCell description]
      [ui/TableCell (values/format-created created)]
+     [ui/TableCell @creator]
      [ui/TableCell (when next-heartbeat-moment (utils/last-time-online next-heartbeat-moment refresh-interval @locale))]
+     [ui/TableCell (or engine-version (str version ".x.x"))]
      [ui/TableCell [uix/Tags tags]]
      [ui/TableCell {:collapsing true}
       (when (some #{id} managers)
@@ -774,14 +780,15 @@
   (let [nuvlaboxes        (subscribe [::subs/nuvlaboxes])
         nuvlabox-clusters (subscribe [::subs/nuvlabox-clusters])
         managers          (distinct
-                           (apply concat
-                                  (map :nuvlabox-managers (:resources @nuvlabox-clusters))))
+                            (apply concat
+                                   (map :nuvlabox-managers (:resources @nuvlabox-clusters))))
         current-cluster   (subscribe [::subs/nuvlabox-cluster])
         selected-nbs      (if @current-cluster
                             (for [target-nb-id (concat (:nuvlabox-managers @current-cluster)
                                                        (:nuvlabox-workers @current-cluster))]
                               (into {} (get (group-by :id (:resources @nuvlaboxes)) target-nb-id)))
                             (:resources @nuvlaboxes))
+        maj-version-only? (subscribe [::subs/one-edge-with-only-major-version (map :id selected-nbs)])
         tr                (subscribe [::i18n-subs/tr])]
     [:div style/center-items
      [ui/Table {:compact "very", :selectable true}
@@ -792,7 +799,11 @@
         [ui/TableHeaderCell "name"]
         [ui/TableHeaderCell "description"]
         [ui/TableHeaderCell (@tr [:created])]
-        [ui/TableHeaderCell (@tr [:last-online])]
+        [ui/TableHeaderCell (@tr [:created-by])]
+        [ui/TableHeaderCell {:single-line true} (@tr [:last-online])]
+        [ui/TableHeaderCell {:single-line true}
+         (@tr [:version])
+         (when @maj-version-only? (ff/help-popup (@tr [:edges-version-info])))]
         [ui/TableHeaderCell "tags"]
         [ui/TableHeaderCell "manager"]]]
 
@@ -822,8 +833,8 @@
   (let [nuvlaboxes        (subscribe [::subs/nuvlaboxes])
         nuvlabox-clusters (subscribe [::subs/nuvlabox-clusters])
         managers          (distinct
-                           (apply concat
-                                  (map :nuvlabox-managers (:resources @nuvlabox-clusters))))
+                            (apply concat
+                                   (map :nuvlabox-managers (:resources @nuvlabox-clusters))))
         selected-nbs      (:resources @nuvlaboxes)]
     [:div style/center-items
      [ui/CardGroup {:centered    true
@@ -863,11 +874,11 @@
                 :style     {:padding-bottom 10}}
        [ui/GridColumn {:width 4}
         [full-text-search-plugin/FullTextSearch
-         {:db-path      [::spec/edges-search]
-          :change-event [::pagination-plugin/change-page
-                         [::spec/pagination] 1]
+         {:db-path            [::spec/edges-search]
+          :change-event       [::pagination-plugin/change-page
+                               [::spec/pagination] 1]
           :placeholder-suffix (str " " @(subscribe [::subs/state-selector]))
-          :style {:width "100%"}}]]
+          :style              {:width "100%"}}]]
        [ui/GridColumn {:width 10}
         (if (= @view-type :cluster)
           [views-clusters/StatisticStates]