+++ code/src/cljs/sixsq/nuvla/ui/profile/views.cljs
@@ -261,14 +261,16 @@
 
 
 (defn TwoFactorTokenInputEnable
-  []
+  [{:keys [show-confirmation]
+    :or {show-confirmation true}}]
   (let [tr (subscribe [::i18n-subs/tr])]
     [TwoFactorTokenInput
      {:on-change #(dispatch [::events/clear-error-message])
       :on-submit #(dispatch [::events/two-factor-auth-callback-exec %1
                              [::events/two-factor-enabled
                               (@tr [:two-factor-auth-enabled])
-                              (@tr [:two-factor-auth-code-accepted-enabled])]])}]))
+                              (@tr [:two-factor-auth-code-accepted-enabled])
+                              show-confirmation]])}]))
 
 
 (defmulti ModalTwoFactorContent identity)
@@ -347,7 +349,7 @@
   [_step]
   [:<>
    [ui/ModalContent
-    [TwoFactorTokenInputEnable]]])
+    [TwoFactorTokenInputEnable {:show-confirmation false}]]])
 
 
 (defn ModalTwoFactorAuth