+++ CHANGELOG.md
@@ -9,8 +9,13 @@
 - acl - Principals dropdown searchable with groups/user
 - Dev tools - Cloudflare preview integration
 - NuvlaEdge overview - Shows last online times for offline edges #895
+- NuvlaEdge overview - Shows engine versions on table view #909
+- NuvlaEdge details - Fixes opening Dropdown menu on key presses inside Update NuvlaEdge form bugfix #908
+- Deployments overview - Shows created by on table and list view #926
+- Edges overview - Shows created by on table and list view #926
+- Profile - Removes auto closing modal on tab change to enable e-mail 2FA bugfix #918
 - Edges and Deployments - Adds filter indicator to full text search
-
+- Filter comp - Show filter in popoup and color the button when filter is set
 
 ## [2.32.7] - 2022-09-29